LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-in-setInterval-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected end of script

PASS window.onerror - compile error in setInterval 
PASS window.onerror - compile error in setInterval (column) 
